tweaks and split up date-picker page into two components
